FoodDelivery/requirements.txt
Django~=4.0.6
djangorestframework~=3.13.1
mysqlclient==2.1.1
pyotp~=2.6.0
PyJWT~=2.4.0
cryptography~=37.0.4
rsa~=4.9
Pillow~=9.2.0
gunicorn==20.1.0
django-environ==0.8.1
django-cors-headers~=3.13.0
requests~=2.28.1
xmltodict~=0.13.0
pytz~=2022.1
phonenumbers~=8.12.57
pandas==1.5.1
opencv-contrib-python-headless~=4.5.5.64
python-dateutil~=2.8.2
django-crontab~=0.7.1
bcrypt~=4.0.1
redis~=4.4.0
celery~=5.2.7
fcm-django~=1.0.12
firebase-admin~=5.4.0
xhtml2pdf~=0.2.8
pdfkit~=1.0.0
pyfcm~=1.5.4
nh3~=0.2.9